+++ ContentManager.Client/components/home/ResourceHiglight.js
