+++ ContentManager.Client/components/home/ResourceList.js
