+++ ContentManager.Client/components/shared/Footer.js
