+++ ContentManager.Client/components/shared/Layout.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Navbar from 'components/Navbar';
-import Footer from "components/Footer";
+import Navbar from 'components/shared/Navbar';
+import Footer from "components/shared/Footer";
 import { ToastContainer } from "react-toastify";
