+++ ContentManager.Client/components/shared/LoaderSpinner.js
