+++ ContentManager.Client/components/shared/Navbar.js
@@ -1,0 +1,38 @@
+import React from "react";
+import Link from "next/link";
+import { useRouter } from "next/router";
+import ActiveLink from "./ActiveLink";
+import Nav from "react-bootstrap/Nav";
+import Navbar from "react-bootstrap/Navbar";
+import Container from "react-bootstrap/Container";
+import NavDropdown from "react-bootstrap/NavDropdown";
+
+export default function Navigationbar() {
+  const router = useRouter();
+
+  return (
+    <>
+      <Navbar bg="dark" variant="dark" expand="lg">
+        <Container>
+          <Navbar.Brand><Link href="/">Content Manager</Link></Navbar.Brand>
+          <Navbar.Toggle aria-controls="basic-navbar-nav" />
+          <Navbar.Collapse id="basic-navbar-nav">
+            <Nav className="me-auto"></Nav>
+            <Nav className="">
+              <ActiveLink href="/home" type="navLink" exact={true}>Home</ActiveLink>
+              <ActiveLink href="/home/about" type="navLink">About</ActiveLink>
+              <NavDropdown title="Resources" id="resources-nav-dropdown">
+                <ActiveLink href="/resources/add" type="dropdownItem">Add</ActiveLink>
+                <NavDropdown.Item href="#action/3.2">Another action</NavDropdown.Item>
+                <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
+                <NavDropdown.Divider />
+                <NavDropdown.Item href="#action/3.4">Separated link</NavDropdown.Item>
+              </NavDropdown>
+            </Nav>
+          </Navbar.Collapse>
+        </Container>
+      </Navbar>
+    </>
+  );
+};
+
