+++ ContentManager.Client/pages/index.js
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useRouter } from "next/router";
-import LoaderSpinner from "components/LoaderSpinner";
+import LoaderSpinner from "components/shared/LoaderSpinnerSpinner";
 
 export default function Index({ resources }) {
   const router = useRouter();
