+++ ContentManager.Client/pages/resources/details.js
@@ -1,4 +1,4 @@
-import Layout from "components/Layout";
+import Layout from "components/shared/Layout";
 
 export default function DetailsResourcePage() {
